Přidej modely číselníkům odměn
Přidány modely číselníku ocenění. Přidána vazba uděleným oceněním.
Dárci se přiřazují podle rodného čísla, ale nemůže to být klíč.
Neevidujeme dárce jako samostatné záznamy.
